requirements_container.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
numpy
faiss-cpu
sentence-transformers
openai
